rootfs: added a systemd service to autoenroll TPM2 based disk description with systemd-cryptenroll
It's not done in the initramfs because dracut did not add it, and it
would be wiser to just add it there, but no time now. If done so, it
would require extra work to not ask typing the password twice.

Note - it could be made prettier - systemd can still spit some output,
but you can enter the password.

🔐 Please enter current passphrase for disk /dev/vda2: (press TAB for no echo) [  OK  ] Started systemd-udevd.service - Rule-based Manager for Device Events and Files.
[  OK  ] Reached target tpm2.target - Trusted Platform Module.
[   12.320574] e1000 0000:00:03.0 ens3: renamed from eth0
[  OK  ] Listening on systemd-rfkill.socket - Load/Save RF Kill Switch Status /dev/rfkill Watch.
         Mounting proc-sys-fs-binfmt_misc.mount - Arbitrary Executable File Formats File System...
[  OK  ] Mounted proc-sys-fs-binfmt_misc.mount - Arbitrary Executable File Formats File System.
[  OK  ] Finished systemd-binfmt.service - Set Up Additional Binary Formats.