docker: Dockerfile.yocto and building and running scripts (the latter could easily be merged with the other scripts). The idea is to also support mounting of SSTATE_MIRRORS or SSTATE_CACHE and some build materials. The are given the same paths as those I have on my host, merely so I can test them faster
I could merge them, as only one image is required and one depends on the other. The nice thing about the respective Dockerfile.yocto is that it shows the minimum dependecies needed to do both together, and also shows how to construct a testing workflow that is independent, but makes a good basic block.